Added player movement
I added a player movement script and player prefab I commonly used, edited to have a triangular shape (approximating a mole). I tweaked the values of this to work well. I also added a script that points the player in the direction you're facing, which means more VECTOR MATH but it works, has some natural easing, and looks really nice.