Added a scrapdata script
Commented very well, so read the script if you want functionality notes. As for purpose:
This is the scrapdata script I was talking about in the inventory channel - I’ve just added some demo options for types of scrap. When building scrap, please ensure all scrap data is stored in this script, but try not to add new behaviors to it - I want it to be used for inventory items and in-world items.